Fixed Issue with Updating Games, Liking Posts, Adding Friends and Deleting Posts without the Correct User Token

+++ server/controllers/PostController.js
@@ -75,17 +75,24 @@ export const LikeUnlikePost = async(req,res) =>{
     try {
         const {id} = req.params;
         const {UserId} = req.body;
-        const GetPost = await Post.findById(id);
-        const FindLike = GetPost.Likes.get(UserId);
-        if (FindLike){
-            GetPost.Likes.delete(UserId);
+
+        if (req.user.id == UserId){
+        
+            const GetPost = await Post.findById(id);
+            const FindLike = GetPost.Likes.get(UserId);
+            if (FindLike){
+                GetPost.Likes.delete(UserId);
+            }else{
+                GetPost.Likes.set(UserId,true)
+            }
+            const UpdatePost = await Post.findByIdAndUpdate(
+                id,{Likes: GetPost.Likes},{new:true}
+            );
+            res.status(200).json(UpdatePost);
+
         }else{
-            GetPost.Likes.set(UserId,true)
+            res.status(400).json({message: "Unauthorized"})
         }
-        const UpdatePost = await Post.findByIdAndUpdate(
-            id,{Likes: GetPost.Likes},{new:true}
-        );
-        res.status(200).json(UpdatePost);
     } catch(error) {
         // Send response with error message
         res.status(404).json({message: error.message})
@@ -95,10 +102,17 @@ export const LikeUnlikePost = async(req,res) =>{
 // Function to delete a post
 export const DeletePost = async(req,res) =>{
     try {
+        
         const {id} = req.params;
-        const deletedPost = await Post.findByIdAndDelete(id);
-        const Posts = await Post.find();
-        res.status(200).json(Posts);
+        const GetPost = await Post.findById(id);
+        if (req.user.id == GetPost.UserId){
+            const deletedPost = await Post.findByIdAndDelete(id);
+            const Posts = await Post.find();
+            res.status(200).json(Posts);
+        }else{
+            res.status(400).json({message: "Unauthorized"})
+        }
+        
     } catch(error) {
         // Send response with error message
         res.status(404).json({message: error.message})
